feat(issue-28): aplicar fixes de convenciones de código
