+++ src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 import { PublicLayoutComponent } from './layouts/public/public-layout/public-layout.component';
 import { PrivateLayoutComponent } from './layouts/private/private-layout/private-layout.component';
-import { authGuard } from './core/guards/auth.guards';
+import { authGuard } from './core/guards/auth.guard';
 import { publicAuthGuard } from './core/guards/public-auth.guard';
 
 export const routes: Routes = [
@@ -32,27 +32,22 @@ export const routes: Routes = [
       {
         path: 'dashboard',
         loadChildren: () => import('./modules/private/dashboard/dashboard.routes').then(m => m.DASHBOARD_ROUTES),
-        data: { breadcrumb: 'Dashboard' }
       },
       {
         path: 'devices',
         loadChildren: () => import('./modules/private/devices/devices.routes').then(m => m.DEVICES_ROUTES),
-        data: { breadcrumb: 'Devices' }
       },
       {
         path: 'orders',
         loadChildren: () => import('./modules/private/orders/orders.routes').then(m => m.ORDERS_ROUTES),
-        data: { breadcrumb: 'Orders' }
       },
       {
         path: 'employees',
         loadChildren: () => import('./modules/private/employees/employees.routes').then(m => m.EMPLOYEES_ROUTES),
-        data: { breadcrumb: 'Employees' }
       },
       {
         path: 'groups',
         loadChildren: () => import('./modules/private/groups/groups.routes').then(m => m.GROUPS_ROUTES),
-        data: { breadcrumb: 'Groups' }
       }
     ]
   },
