+++ src/app/modules/private/dashboard/models/dashboard-stats.model.ts
