+++ src/app/modules/private/orders/modals/order-delete-modal/order-delete-modal.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, OnInit, EventEmitter, Input, Output } from '@angular/core';
-import { Order } from '../../models/Orders';
+import { Order } from '../../models/orders.model';
 
 /**
  * Componente modal para confirmar la eliminación de una orden. Muestra detalles de la orden y proporciona opciones para confirmar o cancelar la eliminación. Emite eventos para que el componente padre pueda manejar la lógica de eliminación y cierre del modal.
